requirements.txt
django
djangorestframework
django-filter
drf-spectacular
djangorestframework-simplejwt
django-cors-headers